fix dragging bounding box for linear elements
